+++ R/recipes.R
@@ -91,11 +91,11 @@ reproduce <- function(worcs_directory = ".", verbose = TRUE, check_endpoints = T
   has_data <- tryCatch({check_data_resources(dn_worcs = worcs_directory,
                        worcsfile = worcs_file,
                        verbose = verbose)}, error = function(e){
-                         usethis::ui_info("Your analysis might not reproduce because worcs could not identify associated data sources.")
+                         cli_msg("!", "Your analysis might not reproduce because worcs could not identify associated data sources.")
                          return(list(data_original = TRUE))
                        })
   if(any(!has_data$data_original)){
-    col_message(paste0("Your analysis might not reproduce because the following original data sources are not available:\n", paste0(names(has_data$data_files)[!has_data$data_original], collapse = ", ")), success = FALSE)
+    cli_msg("!", paste0("Your analysis might not reproduce because the following original data sources are not available:\n", paste0(names(has_data$data_files)[!has_data$data_original], collapse = ", ")))
   }
 
   if(is.null(worcs_file[["recipe"]])){