+++ R/save_load.R
@@ -98,7 +98,7 @@ closed_data <- function(data,
   Args$save_expression <- substitute(save_expression)
   Args$load_expression <- substitute(load_expression)
   Args[[1L]] <- str2lang("worcs:::save_data")
-  Args[["worcs_directory"]] <- worcs_root(path = worcs_directory)
+  # Args[["worcs_directory"]] <- worcs_root(path = worcs_directory)
   eval(Args, parent.frame())
 }
 
@@ -607,37 +607,56 @@ write_gitig <- function(filename, ..., modify = TRUE){
 #' @param ... Objects of class \code{worcs_data}. The function will check if
 #' these are original or synthetic data.
 #' @param msg Expression containing the message to print in case not all
-#' \code{worcs_data} are original. This message may refer to \code{is_synth},
-#' a logical vector indicating which \code{worcs_data} objects are synthetic.
+#' \code{worcs_data} are original.
+#' @param worcs_directory Character, indicating the WORCS project directory to
+#' which to save data. The default value \code{"."} points to the current
+#' directory.
 #' @return No return value. This function is called for its side effect of
 #' printing a notification message.
 #' @examples
-#' df <- iris
-#' class(df) <- c("worcs_data", class(df))
-#' attr(df, "type") <- "synthetic"
-#' result <- capture.output(notify_synthetic(df, msg = "synthetic"))
+#' if(requireNamespace("withr", quietly = TRUE)){
+#'   withr::with_tempdir({
+#'     file.create(".worcs")
+#'     df <- iris
+#'     class(df) <- c("worcs_data", class(df))
+#'     attr(df, "type") <- "synthetic"
+#'     result <- capture.output(notify_synthetic(df, msg = "it is synthetic"))
+#'     df <- df[1:10, ]
+#'     closed_data(df, codebook = NULL)
+#'     file.remove("df.csv")
+#'     result <- capture.output(notify_synthetic(msg = "synthetic"))
+#'     if(requireNamespace("rmarkdown", quietly = TRUE)){
+#'     add_manuscript(manuscript = "github_document")
+#'     print(readLines("manuscript/manuscript.Rmd"))
+#'     rmarkdown::render("manuscript/manuscript.Rmd")
+#'     #if(!any(grepl("this document is reproduced using synthetic data", readLines("manuscript/manuscript.html")))) stop()
+#'     }
+#'   })
+#' }
 #' @rdname notify_synthetic
 #' @export
 #' @seealso closed_data synthetic add_synthetic
 notify_synthetic <- function(...,
-                             msg = NULL){
+                             msg = NULL,
+                             worcs_directory = "."){
   dots <- list(...)
-  cl <- as.list(match.call()[-1])
-  if(is.null(cl[["msg"]])){
-    msg <- quote(c("**Note that", ifelse(all(is_synth), "all", "some"), "of the data files used to generate this document are synthetic. The original data are not available. Synthetic data can be used to evaluate the reproducibility of the analysis code, but the results should not be substantively interpreted, and will likely deviate from the results generated using the original data. Please contact the authors for more information.**"))
-  }
-  msg <- substitute(msg)
-  if(length(dots) > 0){
+  if(isFALSE(length(dots) > 0)){
+    dn_worcs <- worcs_root(worcs_directory)
+    has_data <- check_data_resources(dn_worcs = dn_worcs,
+                                               worcsfile = yaml::read_yaml(file.path(dn_worcs, ".worcs")),
+                                               verbose = FALSE)
+    is_synth <- !has_data$data_original
+  } else {
     if(!all(sapply(dots, inherits, what = "worcs_data"))){
       stop("Some arguments provided to 'notify_synthetic()' are not objects of class 'worcs_data'.", call. = FALSE)
     }
     is_synth <- sapply(dots, attr, which = "type") == "synthetic"
-  } else {
-    worcs_data <- Filter(function(x) inherits(get(x), "worcs_data"), ls(name = parent.env(environment())))
-    is_synth <- sapply(worcs_data, function(x){ attr(get(x), which = "type") }) == "synthetic"
+  }
+  if(is.null(msg)){
+    msg <- "**Note that this document is reproduced using synthetic data. The original data are not available. Synthetic data can be used to evaluate the reproducibility of the analysis code, but the results should not be substantively interpreted, and will likely deviate from the results generated using the original data. Please contact the authors for more information.**"
   }
   if(any(is_synth)){
-    cat(eval(msg))
+    return(msg)
   }
 }